ooDialog - fix size mismatches on 64-bit, use size_t for argument positions
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@6076 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ extensions/platform/windows/oodialog/APICommon.cpp
@@ -252,7 +252,7 @@ RexxObjectPtr invalidTypeException(RexxThreadContext *c, size_t pos, const char 
     return NULLOBJECT;
 }
 
-void invalidImageException(RexxThreadContext *c, int pos, CSTRING type, CSTRING actual)
+void invalidImageException(RexxThreadContext *c, size_t pos, CSTRING type, CSTRING actual)
 {
     TCHAR buffer[256];
     _snprintf(buffer, sizeof(buffer), "Argument %d must be a %s image; found %s", pos, type, actual);
@@ -463,7 +463,7 @@ void failedToRetrieveException(RexxThreadContext *c, CSTRING item, RexxObjectPtr
     c->RaiseException1(Rexx_Error_Execution_user_defined, c->String(buf));
 }
 
-void nullObjectException(RexxThreadContext *c, CSTRING name, int pos)
+void nullObjectException(RexxThreadContext *c, CSTRING name, size_t pos)
 {
     TCHAR buffer[256];
     if ( pos == 0 )
@@ -608,7 +608,7 @@ bool rxGetUInt32Attribute(RexxMethodContext *context, RexxObjectPtr obj, CSTRING
     return result;
 }
 
-bool requiredClass(RexxThreadContext *c, RexxObjectPtr obj, const char *name, int pos)
+bool requiredClass(RexxThreadContext *c, RexxObjectPtr obj, const char *name, size_t pos)
 {
     if ( obj == NULLOBJECT || ! c->IsOfType(obj, name) )
     {
